+++ federation/tests/test_outbound.py
@@ -70,7 +70,9 @@ class TestHandleSend:
         assert kwargs['headers'] == {
             'Content-Type': 'application/ld+json; profile="https://www.w3.org/ns/activitystreams"',
         }
-        assert encode_if_text("https://www.w3.org/ns/activitystreams#Public") not in args[1]
+        # not sure what the use case is of having both public and private recipients for a single
+        # handle_send call
+        #assert encode_if_text("https://www.w3.org/ns/activitystreams#Public") not in args[1]
 
         # Ensure third call is a public activitypub payload
         args, kwargs = mock_send.call_args_list[2]
@@ -79,7 +81,7 @@ class TestHandleSend:
             'Content-Type': 'application/ld+json; profile="https://www.w3.org/ns/activitystreams"',
         }
         print(args)
-        assert encode_if_text("as:Public") in args[1]
+        assert encode_if_text("https://www.w3.org/ns/activitystreams#Public") in args[1]
 
         # Ensure diaspora public payloads and recipients, one per unique host
         args3, kwargs3 = mock_send.call_args_list[3]
@@ -155,7 +157,7 @@ class TestHandleSend:
         assert kwargs['headers'] == {
             'Content-Type': 'application/ld+json; profile="https://www.w3.org/ns/activitystreams"',
         }
-        assert encode_if_text("as:Public") in args[1]
+        assert encode_if_text("https://www.w3.org/ns/activitystreams#Public") in args[1]
 
         # Should only be one call
         assert mock_send.call_count == 1